pyproject.toml
[project]
name = "grokking"
version = "0.1.0"
dependencies = [
    "pyyaml>=6.0.1",
    "torch>=2.4.0",
    "wandb>=0.17.5",
    "accelerate>=0.33.0",
    "pydantic>=2.8.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.3.2",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/grokking"]
